requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.2
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
static3==0.7.0
whitenoise==4.1.2
